Phone number error

+++ db/migrate/20141214045814_create_companies.rb
@@ -2,7 +2,7 @@ class CreateCompanies < ActiveRecord::Migration
   def change
     create_table :companies do |t|
       t.string :name
-      t.integer :phone
+      t.decimal :phone
       t.string :city
       t.string :state
       t.string :country
